core_v1 - continue
regfile read+alu turned out too slow so I added an execution
stage for the sole purpose of registering an output of RF read.
After that it turned out that tcm read is also pretty slow so
only very little can be done in decode stage.
I am still fighting with consequences of that slowness.

TODO
- complete LDB/LDBU/LDH/LDHU
- May be, read register A unconditionally?