core_v1 - continue
Redo decode/register read.
Now the only use for unregistered instruction word is for driving
address port of regfile with address of register A.
The rest of decode uses registered instruction word.
Now the timing is almost sutisfactory, the remaining problems are related
to use of shifter for alignment of readdata which is not complete anyway.

TODO
- complete LDB/LDBU/LDH/LDHU