core_v1 - continue
Execute unconditional PC-relative branches (BR) in 4 clock cycles instead
of 6 clocks.
It made core a little bigger, but it likely worth it, because compiler
generates relatively many BR instructions, including in hot spots.
Of course, if I was in control of compiler, all BR could have been
replaced by JMPI, but I am not in  control of compiler.