core_v1 - continue
Make all PC-relative branches, both conditional and unconditional,
one clock faster, by driving combinatorial (instead of registered) program
counter into tcm_rdaddress.
It costs ~3% in area and few MHz in Fmax, but it almost certainly worth
it because PC-relative branches are VERY common.
In particular Fmax that I see in my tiny test design is not representative
of Fmax of real designs with soft cores. The later tends to be much lower
due to parts of design that have nothing to do with soft core.